pyproject.toml
[tool.poetry]
name = "repl_python3_KAHBAM"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^2.0.1"
vc = "^2018.7.10"
youtube-dl = "^2021.6.6"
discord-pretty-help = "^1.3.1"
jishaku = "^2.0.0"
PyNaCl = "^1.4.0"
discord-py-slash-command = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
